move refund reason options to separate file
